+++ Mods ChoGGi/RC Mechanic/Code/Script.lua
@@ -3,7 +3,7 @@
 -- tell people how to get my library mod (if needs be)
 function OnMsg.ModsReloaded()
 	-- version to version check with
-	local min_version = 59
+	local min_version = 60
 	local idx = table.find(ModsLoaded,"id","ChoGGi_Library")
 	local p = Platform